requirements.txt
aiofiles==24.1.0
aiosqlite==0.17.0
annotated-types==0.5.0
anyio==4.0.0
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.29.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0        
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.0.0.post2     
exceptiongroup==1.1.3
fastapi==0.115.5
Flask==3.1.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0      
google-auth-oauthlib==1.2.1      
googleapis-common-protos==1.66.0 
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.0
idna==3.4
iniconfig==2.0.0
iso8601==1.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
oauthlib==3.2.2
packaging==24.1
pluggy==1.5.0
proto-plus==1.25.0
protobuf==5.29.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.18.0
pyparsing==3.2.0
pypika-tortoise==0.1.6
pytest==8.2.2
python-dateutil==2.8.2
python-decouple==3.8
python-multipart==0.0.6
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
starlette==0.41.3
toml==0.10.2
tortoise-orm==0.21.3
typer==0.12.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.23.2
Werkzeug==3.1.3